+++ dmn-incubation-api-quarkus/src/main/java/org/acme/CustomRestDecision.java
@@ -15,6 +15,8 @@
  */
 package org.acme;
 
+import java.util.Map;
+
 import javax.inject.Inject;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.POST;
@@ -22,29 +24,30 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-import org.kie.kogito.incubation.application.*;
-import org.kie.kogito.incubation.common.*;
-import org.kie.kogito.incubation.predictions.PredictionIds;
-import org.kie.kogito.incubation.predictions.services.PredictionService;
+import org.kie.kogito.incubation.application.AppRoot;
+import org.kie.kogito.incubation.common.DataContext;
+import org.kie.kogito.incubation.common.MapDataContext;
+import org.kie.kogito.incubation.decisions.DecisionIds;
+import org.kie.kogito.incubation.decisions.services.DecisionService;
 
-@Path("/custom")
-public class GreetingResource {
+@Path("/custom-rest-decision")
+public class CustomRestDecision {
 
     @Inject
     AppRoot appRoot;
     @Inject
-    PredictionService svc;
+    DecisionService svc;
 
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public DataContext hello(MapDataContext ctx) {
-        // path: /predictions/LinReg        
+    public DataContext trafficViolation(Map<String, Object> payload) {
+        // path: /decisions/https%3A%2F%2Fgithub.com%2Fkiegroup%2Fdrools%2Fkie-dmn%2F_A4BCA8B8-CF08-433F-93B2-A2598F19ECFF/Traffic%20Violation
+
         var id = appRoot
-                .get(PredictionIds.class)
-                .get("LinReg");
-
-        return svc.evaluate(id, ctx);
+                .get(DecisionIds.class)
+                .get("https://github.com/kiegroup/drools/kie-dmn/_A4BCA8B8-CF08-433F-93B2-A2598F19ECFF",
+                        "Traffic Violation");
+        return svc.evaluate(id, MapDataContext.from(payload));
     }
-
 }
