+++ pmml-incubation-api-quarkus/src/main/java/org/acme/CustomRestPrediction.java
@@ -15,8 +15,6 @@
  */
 package org.acme;
 
-import java.util.Map;
-
 import javax.inject.Inject;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.POST;
@@ -24,27 +22,29 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-import org.kie.kogito.incubation.application.AppRoot;
-import org.kie.kogito.incubation.common.DataContext;
-import org.kie.kogito.incubation.common.MapDataContext;
-import org.kie.kogito.incubation.processes.ProcessIds;
-import org.kie.kogito.incubation.processes.services.StraightThroughProcessService;
+import org.kie.kogito.incubation.application.*;
+import org.kie.kogito.incubation.common.*;
+import org.kie.kogito.incubation.predictions.PredictionIds;
+import org.kie.kogito.incubation.predictions.services.PredictionService;
 
-@Path("/hello")
-public class GreetingResource {
+@Path("/custom-rest-prediction")
+public class CustomRestPrediction {
 
     @Inject
     AppRoot appRoot;
     @Inject
-    StraightThroughProcessService svc;
+    PredictionService svc;
 
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public DataContext hello(Map<String, Object> payload) {
-        // path: /processes/scripts
+    public DataContext linearRegression(MapDataContext ctx) {
+        // path: /predictions/LinReg        
+        var id = appRoot
+                .get(PredictionIds.class)
+                .get("LinReg");
 
-        var id = appRoot.get(ProcessIds.class).get("scripts");
-        return svc.evaluate(id, MapDataContext.from(payload));
+        return svc.evaluate(id, ctx);
     }
+
 }
